refactor: 이벤트 핸들러의 위치를 발행자 기준에서 구독자 기준으로 변경 (#1375)
* refactor: 멤버 관련 이벤트 처리 로직 멤버 핸들러로 이동

* refactor: 멤버십 관련 이벤트 처리 로직 멤버십 핸들러로 이동

* refactor: 불필요 TODO 주석 제거
